+++ app/src/main/java/woowacourse/shopping/domain/mapper/CartMapper.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.domain.mapper
+
+import woowacourse.shopping.domain.model.Cart
+import woowacourse.shopping.presentation.ui.model.CartModel
+
+fun Cart.toUiModel(isChecked: Boolean = false) =
+    CartModel(cartId, product.id, product.name, product.imgUrl, product.price, quantity, isChecked, calculatedPrice)
