+++ app/src/main/java/woowacourse/shopping/domain/mapper/CouponMapper.kt
@@ -1,0 +1,16 @@
+package woowacourse.shopping.domain.mapper
+
+import woowacourse.shopping.domain.model.Coupon
+import woowacourse.shopping.presentation.ui.model.CouponModel
+import java.time.format.DateTimeFormatter
+
+fun Coupon.toUiModel(formatter: DateTimeFormatter): CouponModel {
+    return CouponModel(
+        id = id.toLong(),
+        description = description,
+        expiredDate = expirationDate.format(formatter),
+        minimumAmount = minimumOrderedAmount,
+        discountAmount = discountAmount,
+        isChecked = false,
+    )
+}
