+++ app/src/main/java/woowacourse/shopping/domain/mapper/ProductMapper.kt
@@ -1,0 +1,6 @@
+package woowacourse.shopping.domain.mapper
+
+import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.presentation.ui.model.ProductModel
+
+fun Product.toUiModel(quantity: Int = 0) = ProductModel(id, name, imgUrl, price, category, quantity)
