+++ app/src/main/java/woowacourse/shopping/presentation/ui/model/CartModel.kt
@@ -1,7 +1,5 @@
 package woowacourse.shopping.presentation.ui.model
 
-import woowacourse.shopping.domain.model.Cart
-
 data class CartModel(
     val cartId: Long,
     val productId: Long,
@@ -12,6 +10,3 @@ data class CartModel(
     val isChecked: Boolean,
     val calculatedPrice: Int,
 )
-
-fun Cart.toUiModel(isChecked: Boolean = false) =
-    CartModel(cartId, product.id, product.name, product.imgUrl, product.price, quantity, isChecked, calculatedPrice)
