+++ app/src/main/java/woowacourse/shopping/presentation/ui/model/ProductModel.kt
@@ -1,7 +1,5 @@
 package woowacourse.shopping.presentation.ui.model
 
-import woowacourse.shopping.domain.model.Product
-
 data class ProductModel(
     val id: Long,
     val name: String,
@@ -17,5 +15,3 @@ data class ProductModel(
         val INVALID_PRODUCT_MODEL = ProductModel(id = -1L, "", "", 0L, "", 0)
     }
 }
-
-fun Product.toUiModel(quantity: Int = 0) = ProductModel(id, name, imgUrl, price, category, quantity)
